+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changes
 
 ### Version 10.1.0
-- Adds configuration `bootstrap.mdc.includeRequest` to include the request in MDC for logs. E.g. `GET MyController.method`
+- Adds configuration `bootstrap.mdc.includeHandler` to include the request in MDC for logs. E.g. `GET MyController.method`
 
 ### Version 10.0.0
 - MDC Changes.